# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Notebook API",
        "contact": {
            "email": "admin@sartoj.tech"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Notebook API server"
        }
    ],
    "paths": {
        "/notebook": {
            "get": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Display all Notebooks. 10 items per page",
                "operationId": "notebookAll",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number.",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "4"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Notebook not found"
                    },
                    "405": {
                        "description": "Method not allowed"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "503": {
                        "description": "Service unavailable"
                    }
                }
            },
            "post": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Create yet another notebook",
                "operationId": "notebookCreate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NotebookStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Note created successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Notebook not found"
                    },
                    "405": {
                        "description": "Method not allowed"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "503": {
                        "description": "Service unavailable"
                    }
                }
            }
        },
        "/notebook/{id}": {
            "get": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Display Notebook by ID",
                "operationId": "notebookByID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of the Notebook",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notebook founded successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Notebook not found"
                    },
                    "405": {
                        "description": "Method not allowed"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "503": {
                        "description": "Service unavailable"
                    }
                }
            },
            "put": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Create yet another notebook",
                "operationId": "notebookUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of the Notebook",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NotebookStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Note updated successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Notebook not found"
                    },
                    "405": {
                        "description": "Method not allowed"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "503": {
                        "description": "Service unavailable"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Delete Notebook by ID",
                "operationId": "deleteByID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id of the Notebook",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "5"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Note deleted"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Notebook not found"
                    },
                    "405": {
                        "description": "Method not allowed"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "503": {
                        "description": "Service unavailable"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "NotebookStoreRequest": {
                "title": "Notebook showing request",
                "description": "Creating Notebook",
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Full name",
                        "type": "string",
                        "example": "John Doe Doe"
                    },
                    "company": {
                        "title": "company",
                        "description": "Company name",
                        "type": "string",
                        "example": "SARTOJ Tech."
                    },
                    "phone": {
                        "title": "phone",
                        "description": "Your phone number",
                        "type": "string",
                        "example": "+7777777777"
                    },
                    "email": {
                        "title": "email",
                        "description": "Your email",
                        "type": "string",
                        "example": "info@sartoj.tech"
                    },
                    "birthday": {
                        "title": "birthday",
                        "description": "Your bitrthday",
                        "type": "string",
                        "format": "date",
                        "example": "2022-02-22"
                    },
                    "photo": {
                        "title": "photo",
                        "description": "Your photo url",
                        "type": "string",
                        "example": "avatar.png"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Notebook",
            "description": "To get notebooks from database"
        }
    ]
}